extract load_from_raw_data
